# openapi.yaml
openapi: 3.1.0
info:
  title: Grounded & Fruitful API
  version: 1.0.0
  description: >-
    Authorization uses canonical roles from active memberships plus organization,
    program, and resource relationships. Custom claims and route entry never grant
    tenant resource access. Domain Firestore access is server-only.
servers: [{ url: /api/v1 }]
...
paths:
...
  /admin/bible-content/imports:
    get:
      tags: [Bible]
      operationId: listBibleImports
      description: Lists tenant-scoped import summaries with server-derived review actions.
      security: [{ firebaseBearer: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: quarterId, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string, maxLength: 100 } }
        - { name: cursor, in: query, schema: { type: string } }
        - {
            name: limit,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 50, default: 25 },
          }
      responses:
        { "200": { description: Bounded import summaries and next cursor } }
    post:
      tags: [Bible]
      operationId: createBibleImport
      description: Admin-only two-file preview import. Parsing never publishes content.
      security: [{ firebaseBearer: [] }]
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Stable 8-128 character key. An exact tenant-scoped retry returns the existing import without uploading again.
          schema: { type: string, pattern: "^[A-Za-z0-9_-]{8,128}$" }
      requestBody:
        required: true
        description: Use the text parts `organizationId`, `quarterId`, and `title`, and exactly one DOCX file part each named `quizFile` and `answerKeyFile`.
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/BibleImportUpload" }
            encoding:
              organizationId: { contentType: text/plain }
              quarterId: { contentType: text/plain }
              title: { contentType: text/plain }
              quizFile:
                {
                  contentType: application/vnd.openxmlformats-officedocument.wordprocessingml.document,
                }
              answerKeyFile:
                {
                  contentType: application/vnd.openxmlformats-officedocument.wordprocessingml.document,
                }
      responses:
        {
          "201": { description: Draft preview },
          "403": { $ref: "#/components/responses/Forbidden" },
          "404": { $ref: "#/components/responses/NotFound" },
          "409": { $ref: "#/components/responses/Conflict" },
          "415": { $ref: "#/components/responses/Validation" },
          "422": { $ref: "#/components/responses/Validation" },
          "500": { $ref: "#/components/responses/Internal" },
          "503": { $ref: "#/components/responses/Internal" },
        }
  /admin/bible-content/imports/{importId}:
    get:
      tags: [Bible]
      operationId: getBibleImport
      security: [{ firebaseBearer: [] }]
      parameters: [{ $ref: "#/components/parameters/ImportId" }]
      responses:
        {
          "200": { description: Tenant-scoped safe preview },
          "404": { $ref: "#/components/responses/NotFound" },
        }
  /admin/bible-content/imports/{importId}/items/{itemId}:
    patch:
      tags: [Bible]
      operationId: correctBibleImportItem
      description: Corrects preview only; expectedVersion provides optimistic concurrency.
      security: [{ firebaseBearer: [] }]
      parameters:
        [
          { $ref: "#/components/parameters/ImportId" },
          { name: itemId, in: path, required: true, schema: { type: string } },
        ]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/BibleItemPatch" } },
            },
        }
      responses:
        {
          "200": { description: Corrected preview },
          "409": { $ref: "#/components/responses/Conflict" },
        }
  /admin/bible-content/imports/{importId}/validate:
    post:
      tags: [Bible]
      operationId: validateBibleImport
      security: [{ firebaseBearer: [] }]
      parameters: [{ $ref: "#/components/parameters/ImportId" }]
      responses: { "200": { description: Validation errors and warnings } }
  /admin/bible-content/imports/{importId}/commit:
    post:
      tags: [Bible]
      operationId: commitBibleImport
      description: Idempotently commits validated preview as draft; never publishes.
      security: [{ firebaseBearer: [] }]
      parameters: [{ $ref: "#/components/parameters/ImportId" }]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/BibleLifecycle" } },
            },
        }
      responses:
        {
          "200": { description: Draft content set identifier },
          "409": { $ref: "#/components/responses/Conflict" },
        }
  /admin/bible-content/imports/{importId}/reprocess:
    post:
      tags: [Bible]
      operationId: reprocessBibleImport
      security: [{ firebaseBearer: [] }]
      parameters: [{ $ref: "#/components/parameters/ImportId" }]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/BibleLifecycle" } },
            },
        }
      responses: { "200": { description: Processing import } }
  /admin/bible-content/imports/{importId}/reject:
    post:
      tags: [Bible]
      operationId: rejectBibleImport
      security: [{ firebaseBearer: [] }]
      parameters: [{ $ref: "#/components/parameters/ImportId" }]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/BibleLifecycle" } },
            },
        }
      responses: { "200": { description: Rejected import } }
  /admin/bible-content/imports/{importId}/documents/{kind}:
    get:
      tags: [Bible]
      operationId: downloadBibleImportDocument
      description: Streams a private source document after tenant and reviewer authorization. Answer-key access is audited.
      security: [{ firebaseBearer: [] }]
      parameters:
        - { $ref: "#/components/parameters/ImportId" }
        - {
            name: kind,
            in: path,
            required: true,
            schema: { type: string, enum: [question, answer-key] },
          }
      responses:
        {
          "200": { description: Private DOCX attachment },
          "404": { $ref: "#/components/responses/NotFound" },
        }
  /admin/bible-content:
    get:
      {
        tags: [Bible],
        operationId: listBibleContent,
        security: [{ firebaseBearer: [] }],
        responses: { "200": { description: Tenant-scoped content list } },
      }
  /admin/bible-content/{contentSetId}:
    get:
      {
        tags: [Bible],
        operationId: getBibleContent,
        security: [{ firebaseBearer: [] }],
        parameters: [{ $ref: "#/components/parameters/ContentSetId" }],
        responses:
          {
            "200":
              {
                description: Administrative content including protected answers,
              },
          },
      }
    patch:
      tags: [Bible]
      operationId: updateBibleContent
      security: [{ firebaseBearer: [] }]
      parameters: [{ $ref: "#/components/parameters/ContentSetId" }]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/BibleLifecycle" } },
            },
        }
      responses:
        {
          "200": { description: Updated draft },
          "409": { $ref: "#/components/responses/Conflict" },
        }
  /admin/bible-content/{contentSetId}/publish:
    post:
      {
        tags: [Bible],
        operationId: publishBibleContent,
        security: [{ firebaseBearer: [] }],
        parameters: [{ $ref: "#/components/parameters/ContentSetId" }],
        requestBody:
          {
            required: true,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/BibleLifecycle" } },
              },
          },
        responses:
          {
            "200": { description: Published content },
            "409": { $ref: "#/components/responses/Conflict" },
          },
      }
  /admin/bible-content/{contentSetId}/archive:
    post:
      {
        tags: [Bible],
        operationId: archiveBibleContent,
        security: [{ firebaseBearer: [] }],
        parameters: [{ $ref: "#/components/parameters/ContentSetId" }],
        requestBody:
          {
            required: true,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/BibleLifecycle" } },
              },
          },
        responses: { "200": { description: Archived content } },
      }
...
components:
  parameters:
...
    ImportId:
      {
        name: importId,
        in: path,
        required: true,
        schema: { type: string, minLength: 1, maxLength: 128 },
      }
    ContentSetId:
      {
        name: contentSetId,
        in: path,
        required: true,
        schema: { type: string, minLength: 1, maxLength: 128 },
      }
...
  securitySchemes:
    {
      firebaseBearer:
        { type: http, scheme: bearer, bearerFormat: Firebase ID token },
    }
  responses:
...
    Forbidden:
      {
        description: Resource access denied,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/Error" } },
          },
      }
    Conflict: { description: Version, lifecycle, or idempotency conflict }
    NotFound:
      {
        description: Not found,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/Error" } },
          },
      }
    Validation:
      {
        description: Validation failed,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/Error" } },
          },
      }
...
    Internal:
      {
        description: Unexpected server error,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/Error" } },
          },
      }
  schemas:
...
    BibleImportUpload:
      type: object
      additionalProperties: false
      required: [organizationId, quarterId, title, quizFile, answerKeyFile]
      properties:
        organizationId: { type: string, minLength: 1, maxLength: 128 }
        quarterId: { type: string, minLength: 1, maxLength: 128 }
        title: { type: string, minLength: 1, maxLength: 200 }
        quizFile:
          {
            type: string,
            format: binary,
            description: DOCX quiz document; maximum 5 MiB,
          }
        answerKeyFile:
          {
            type: string,
            format: binary,
            description: DOCX answer-key document; maximum 5 MiB,
          }
    BibleLifecycle:
      type: object
      required: [expectedVersion]
      properties:
        {
          expectedVersion: { type: integer, minimum: 1 },
          acknowledgeWarnings: { type: boolean },
          title: { type: string },
          description: { type: string },
        }
    BibleItemPatch:
      type: object
      required: [expectedVersion]
      properties:
        {
          expectedVersion: { type: integer, minimum: 1 },
          title: { type: string },
          scriptureReference: { type: string },
        }
...
    Error:
      {
        type: object,
        required: [error],
        properties:
          {
            error:
              {
                type: object,
                required: [code, message, requestId],
                properties:
                  {
                    code: { type: string },
                    message: { type: string },
                    requestId: { type: string },
                    fieldErrors: { type: object, additionalProperties: true },
                  },
              },
          },
      }
...